Decrypt files even if filenames map is missing
In case something went wrong during encryption, we might
still be able to salvage the files despite the filenames
map file (which maps encrypted file names to real paths)
not being found. We suffix the files with .clear to
indicate the files are unencrypted and we do a best
effort to decrypt them.

Fixes #1